api/requirements.txt
flask==3.1.1
flask-cors==5.0.1
flask-limiter==3.12
python-dotenv==1.1.0
gunicorn==23.0.0
